Fixing graphics
Long story short, I lost the PSDs I was working with (I only had outdated versions after I reinstalled my Mac Pro in December to unsuccessfully try to fix a bluetooth issue), but I could recover the baked PNGs. So I've done that and fixed the positioning. It works, except on smaller screens (like this MacBook Air I'm on, where the shotcuts bar is covered a bit by the graphics).